backend/pyproject.toml
[project]
name = "guitar-player-backend"
version = "0.1.0"
description = "FastAPI backend for Guitar Player"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "mangum>=0.19.0",
    "uvicorn[standard]>=0.32.0",
    "awslambdaric>=2.0.0",
    "sqlalchemy[asyncio]>=2.0.0",
    "asyncpg>=0.30.0",
    "alembic>=1.14.0",
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "boto3>=1.35.0",
    "python-jose[cryptography]>=3.3.0",
    "yt-dlp>=2026.03.03",
    "bgutil-ytdlp-pot-provider>=1.2.2",
    "httpx>=0.27.0",
    "psycopg2-binary>=2.9.11",
    "email-validator>=2.0.0",
    "openai>=1.0.0",
    "requests>=2.32.5",
    "python-json-logger>=3.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-timeout>=2.2.0",
    "aiosqlite>=0.20.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/guitar_player"]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.pyright]
venvPath = "."
venv = ".venv"
pythonVersion = "3.13"
extraPaths = ["src"]
